bug fixes + cleaning
