+++ src/abms/AbstractSocialABM.jl
@@ -3,6 +3,9 @@ A type specifying special functionalities for elemantary Social ABMs,
     i.e. stuffs that don't exist in Agents.jl  
 """
 
+export AbstractSocialABM
+export addproperty!
+
 "Social ABM to be enhanced with relevant functionlities"
 abstract type AbstractSocialABM <: AbstractABM end 
 
@@ -14,3 +17,17 @@ function addproperty!(model::AbstractSocialABM,property::Symbol,val)
     end 
     model.properties[property] = val  
 end 
+
+
+#=
+It is thinkable to associate further attributes to SocialABMs s.a.
+
+variable(sabm::AbstractSocialABM,var::Symbol) = sabm.variable[var]
+parameter(sabm::AbstractSocialABM,par::Parameter) = sabm.parameter[par]
+data(sabm::AbstractSocialABM,symbol::Symbol)  = sabm.data[symbol]
+... 
+function addData!(sabm,symbol,csvfile) end
+function addVariable!(sabm,symbol) end 
+function deleteVariable!(sabm,symbol) end
+
+=#
